combine contract helpers to ContractFacade
