+++ packages/broker/src/plugins/operator/inspectRandomNode.ts
@@ -1,9 +1,9 @@
 import { EthereumAddress, Logger } from '@streamr/utils'
-import { InspectRandomNodeHelper } from './InspectRandomNodeHelper'
 import { StreamPartAssignments } from './StreamPartAssignments'
 import { StreamrClient } from 'streamr-client'
 import { StreamPartIDUtils } from '@streamr/protocol'
 import { findNodesForTarget, findTarget, inspectTarget } from './inspectionUtils'
+import { ContractFacade } from './ContractFacade'
 
 const logger = new Logger(module)
 
@@ -13,7 +13,7 @@ export type InspectTargetFn = typeof inspectTarget
 
 export async function inspectRandomNode(
     operatorContractAddress: EthereumAddress,
-    helper: InspectRandomNodeHelper,
+    contractFacade: ContractFacade,
     assignments: StreamPartAssignments,
     streamrClient: StreamrClient,
     heartbeatTimeoutInMs: number,
@@ -25,7 +25,7 @@ export async function inspectRandomNode(
 ): Promise<void> {
     logger.info('Select a random operator to inspect')
 
-    const target = await findTargetFn(operatorContractAddress, helper, assignments)
+    const target = await findTargetFn(operatorContractAddress, contractFacade, assignments)
     if (target === undefined) {
         return
     }
@@ -47,7 +47,7 @@ export async function inspectRandomNode(
 
     if (!pass) {
         logger.info('Raise flag', { target })
-        await helper.flag(
+        await contractFacade.flag(
             target.sponsorshipAddress,
             target.operatorAddress,
             StreamPartIDUtils.getStreamPartition(target.streamPart)
