+++ packages/broker/src/plugins/operator/inspectionUtils.ts
@@ -5,11 +5,11 @@ import { StreamID, StreamPartID, StreamPartIDUtils } from '@streamr/protocol'
 import { EthereumAddress, Logger, wait } from '@streamr/utils'
 import { ConsistentHashRing } from './ConsistentHashRing'
 import { StreamPartAssignments } from './StreamPartAssignments'
-import { InspectRandomNodeHelper } from './InspectRandomNodeHelper'
 import { weightedSample } from '../../helpers/weightedSample'
 import sample from 'lodash/sample'
 import without from 'lodash/without'
 import { formCoordinationStreamId } from './formCoordinationStreamId'
+import { ContractFacade } from './ContractFacade'
 
 const logger = new Logger(module)
 
@@ -41,11 +41,11 @@ function getPartitionsOfStreamAssignedToMe(
 
 export async function findTarget(
     myOperatorContractAddress: EthereumAddress,
-    helper: InspectRandomNodeHelper,
+    contractFacade: ContractFacade,
     assignments: StreamPartAssignments
 ): Promise<Target | undefined> {
     // choose sponsorship
-    const sponsorships = await helper.getSponsorshipsOfOperator(myOperatorContractAddress)
+    const sponsorships = await contractFacade.getSponsorshipsOfOperator(myOperatorContractAddress)
     const suitableSponsorships = sponsorships
         .filter(({ operatorCount }) => operatorCount >= 2)  // exclude sponsorships with only self
         .filter(({ streamId }) => isAnyPartitionOfStreamAssignedToMe(assignments, streamId))
@@ -59,7 +59,7 @@ export async function findTarget(
     )!
 
     // choose operator
-    const operators = await helper.getOperatorsInSponsorship(targetSponsorship.sponsorshipAddress)
+    const operators = await contractFacade.getOperatorsInSponsorship(targetSponsorship.sponsorshipAddress)
     const targetOperatorAddress = sample(without(operators, myOperatorContractAddress))
     if (targetOperatorAddress === undefined) {
         // Only happens if during the async awaits the other operator(s) were removed from the sponsorship.
