+++ packages/broker/src/plugins/operator/maintainOperatorValue.ts
@@ -1,20 +1,20 @@
 import { Logger } from '@streamr/utils'
-import { MaintainOperatorValueHelper } from './MaintainOperatorValueHelper'
 import { multiply } from '../../helpers/multiply'
+import { ContractFacade } from './ContractFacade'
 
 const logger = new Logger(module)
 
 export const maintainOperatorValue = async (
     withdrawLimitSafetyFraction: number,
-    helper: MaintainOperatorValueHelper
+    contractFacade: ContractFacade
 ): Promise<void> => {
     logger.info('Check whether it is time to withdraw my earnings')
-    const { sumDataWei, maxAllowedEarningsDataWei, sponsorshipAddresses } = await helper.getMyEarnings()
+    const { sumDataWei, maxAllowedEarningsDataWei, sponsorshipAddresses } = await contractFacade.getMyEarnings()
     const triggerWithdrawLimitDataWei = multiply(maxAllowedEarningsDataWei, 1 - withdrawLimitSafetyFraction)
     logger.trace(` -> is ${sumDataWei} > ${triggerWithdrawLimitDataWei} ?`)
     if (sumDataWei > triggerWithdrawLimitDataWei) {
         logger.info('Withdraw earnings from sponsorships', { sponsorshipAddresses })
-        await helper.withdrawMyEarningsFromSponsorships(sponsorshipAddresses)
+        await contractFacade.withdrawMyEarningsFromSponsorships(sponsorshipAddresses)
     } else {
         logger.info('Skip withdrawing earnings')
     }
